Move CompleteGorcSystem and related types to system.rs
Refactored the GORC system by moving CompleteGorcSystem, GorcPerformanceReport, GORC_VERSION, and MAX_CHANNELS from mod.rs to a new system.rs module. Updated mod.rs to re-export these items from system.rs for cleaner organization and separation of concerns.
